Alter scheulder to run each date cycle per min, rather than as soon as possible to reduce instant load / mem usage from xml parsing
